+++ core/src/filed/backup.cc
@@ -1177,7 +1177,7 @@ bool EncodeAndSendAttributes(JobControlRecord* jcr,
 
   Dmsg1(300, "encode_and_send_attrs fname=%s\n", ff_pkt->fname);
   /** Find what data stream we will use, then encode the attributes */
-  if ((data_stream = SelectDataStream(ff_pkt, me->compatible)) == STREAM_NONE) {
+  if ((data_stream = SelectDataStream(ff_pkt)) == STREAM_NONE) {
     /* This should not happen */
     Jmsg0(jcr, M_FATAL, 0,
           _("Invalid file flags, no supported data stream type.\n"));