+++ core/src/findlib/attribs.h
@@ -1,7 +1,7 @@
 /*
    BAREOS® - Backup Archiving REcovery Open Sourced
 
-   Copyright (C) 2018-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2018-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -27,6 +27,6 @@ int encode_attribsEx(JobControlRecord* jcr,
 bool SetAttributes(JobControlRecord* jcr,
                    Attributes* attr,
                    BareosFilePacket* ofd);
-int SelectDataStream(FindFilesPacket* ff_pkt, bool compatible);
+int SelectDataStream(FindFilesPacket* ff_pkt);
 
 #endif  // BAREOS_FINDLIB_ATTRIBS_H_