+++ core/src/stored/ansi_label.cc
@@ -3,7 +3,7 @@
 
    Copyright (C) 2005-2009 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -63,10 +63,8 @@ int ReadAnsiIbmLabel(DeviceControlRecord* dcr)
   char* VolName = dcr->VolumeName;
   bool ok = false;
 
-  /*
-   * Read VOL1, HDR1, HDR2 labels, but ignore the data
-   * If tape read the following EOF mark, on disk do not read.
-   */
+  /* Read VOL1, HDR1, HDR2 labels, but ignore the data
+   * If tape read the following EOF mark, on disk do not read. */
   Dmsg0(100, "Read ansi label.\n");
   if (!dcr->dev->IsTape()) { return VOL_OK; }
 
@@ -163,27 +161,13 @@ int ReadAnsiIbmLabel(DeviceControlRecord* dcr)
           return VOL_LABEL_ERROR;
         }
 
-        if (me->compatible) {
-          if (!bstrncmp("BACULA.DATA", &label[4], 11)
-              && !bstrncmp("BAREOS.DATA", &label[4], 11)) {
-            Dmsg1(100,
-                  "HD1 not Bacula/Bareos label. Wanted BACULA.DATA/BAREOS.DATA "
-                  "got %11s\n",
-                  &label[4]);
-            Mmsg1(jcr->errmsg,
-                  _("ANSI/IBM Volume \"%s\" does not belong to Bareos.\n"),
-                  dcr->dev->VolHdr.VolumeName);
-            return VOL_NAME_ERROR; /* Not a Bareos label */
-          }
-        } else {
-          if (!bstrncmp("BAREOS.DATA", &label[4], 11)) {
-            Dmsg1(100, "HD1 not Bareos label. Wanted BAREOS.DATA got %11s\n",
-                  &label[4]);
-            Mmsg1(jcr->errmsg,
-                  _("ANSI/IBM Volume \"%s\" does not belong to Bareos.\n"),
-                  dcr->dev->VolHdr.VolumeName);
-            return VOL_NAME_ERROR; /* Not a Bareos label */
-          }
+        if (!bstrncmp("BAREOS.DATA", &label[4], 11)) {
+          Dmsg1(100, "HD1 not Bareos label. Wanted BAREOS.DATA got %11s\n",
+                &label[4]);
+          Mmsg1(jcr->errmsg,
+                _("ANSI/IBM Volume \"%s\" does not belong to Bareos.\n"),
+                dcr->dev->VolHdr.VolumeName);
+          return VOL_NAME_ERROR; /* Not a Bareos label */
         }
         Dmsg0(100, "Got HDR1 label\n");
         break;
@@ -299,10 +283,8 @@ bool WriteAnsiIbmLabels(DeviceControlRecord* dcr, int type, const char* VolName)
   time_t now;
   int len, status, label_type;
 
-  /*
-   * If the Device requires a specific label type use it,
-   * otherwise, use the type requested by the Director
-   */
+  /* If the Device requires a specific label type use it,
+   * otherwise, use the type requested by the Director */
   if (dcr->device_resource->label_type != B_BAREOS_LABEL) {
     label_type = dcr->device_resource->label_type; /* force label type */
   } else {
@@ -324,10 +306,8 @@ bool WriteAnsiIbmLabels(DeviceControlRecord* dcr, int type, const char* VolName)
         return false;
       }
 
-      /*
-       * ANSI labels have 6 characters, and are padded with spaces 'vol1\0' =>
-       * 'vol1   \0'
-       */
+      /* ANSI labels have 6 characters, and are padded with spaces 'vol1\0' =>
+       * 'vol1   \0' */
       strcpy(ansi_volname, VolName);
       for (int i = len; i < 6; i++) { ansi_volname[i] = ' '; }
       ansi_volname[6] = '\0'; /* only for debug */
@@ -360,11 +340,8 @@ bool WriteAnsiIbmLabels(DeviceControlRecord* dcr, int type, const char* VolName)
       SerBegin(label, sizeof(label));
       SerBytes(labels[type], 3);
       SerBytes("1", 1);
-      if (me->compatible) {
-        SerBytes("BACULA.DATA", 11); /* Filename field */
-      } else {
-        SerBytes("BAREOS.DATA", 11); /* Filename field */
-      }
+      SerBytes("BAREOS.DATA", 11); /* Filename field */
+
       SerBegin(&label[21], sizeof(label) - 21); /* fileset field */
       SerBytes(ansi_volname, 6);                /* write Vol Ser No. */
       SerBegin(&label[27], sizeof(label) - 27);