+++ core/src/stored/sd_plugins.h
@@ -3,7 +3,7 @@
 
    Copyright (C) 2007-2011 Free Software Foundation Europe e.V.
    Copyright (C) 2011-2012 Planets Communications B.V.
-   Copyright (C) 2013-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -68,8 +68,7 @@ typedef enum
   bsdVarJobErrors = 13,
   bsdVarJobFiles = 14,
   bsdVarJobBytes = 15,
-  bsdVarCompatible = 16,
-  bsdVarPluginDir = 17
+  bsdVarPluginDir = 16
 } bsdrVariable;
 
 // Bareos Variable Ids (Write)