+++ core/src/win32/compat/include/compat.h
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2004-2011 Free Software Foundation Europe e.V.
-   Copyright (C) 2019-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2019-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -346,9 +346,6 @@ char* win32_cgets(char* buffer, int len);
 
 int WSA_Init(void);
 void Win32TSDCleanup();
-void Win32ClearCompatible();
-void Win32SetCompatible();
-bool Win32IsCompatible();
 
 #if defined(HAVE_MINGW)
 void closelog();